Fix grammar
Fix error in grammar: "A lexeme in G1 is not a lexeme in any of the lexers:
ndl_day" (or ndl_week, ndl_hour, etc).

From Jeffrey Kegler's explanation in marpa-parser list: "G1 thinks that <name>
is a lexeme. L0 disagrees. In more detail, G1 finds that <name> is on the RHS of
one of its rules, and not on a LHS. So G1 concludes that <name> is a lexeme. L0
looks for <name> on the *LHS* of one of its rules. It does find it on the LHS of
an L0 rule, as required. But it also finds this: 'names ~ name+'. Here <name>
appears on the RHS of an L0 rule, which tells Marpa that is it not a lexeme,
after all. One solution (if my guess is right) is to rewrite so that <name> is
not on the RHS of any L0 rules."

So basically, I have ndl_{year,month,day,...} on the left side of L0 rule as
well as right side of L0 rule (so it's not a lexeme). The solution is indeed to
rewrite so that they are not on the RHS of any L0 rules.